don't have source system/component that conflicts with GCS
Previously was using system=255 but that was the same as the ground
station. Thus, when it completed downloading parameters, none of our
landing target messages would be sent to the autopilot. This is fixed if
we set system=1 (i.e. RPi is physically on the drone) and component=2 to
not conflict with anything.